Use `captionsUrl` from data rather than option
Saves users having to make a separate request to lookup the captionsUrl.

This is a breaking change for users (and tests) calling `addVideo()`
directly. Some of the tests were doing this and modifying `videoData`
internally, whereas we now need `getData()` to have run in order to
get the `captionsUrl`. An error will be thrown if the situation arises.

Introduces a new option:
  `showCaptions` - true by default

Removes:
  `captionsUrl` - this now comes from the data (by API or attribute)