Dispatch progress events only once per pecentage (#138)
Progress percentage is rounded so that 1 percent of a video (especially
a long video) may take many seconds. During this time the tracking
will fire everytime. When a video is stalling the progress event if
fired many times resulting in many requests which were impacting the
loading of the video further.